refactor: extract fetchSettings method
